#21 썸네일 이미지만 압축
