+++ src/main/java/com/artfolio/artfolio/config/RedisConfig.java
@@ -17,5 +17,6 @@ public class RedisConfig {
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
         return new LettuceConnectionFactory(host, port);
+
     }
 }
