+++ src/main/java/com/artfolio/artfolio/controller/RealTimeAuctionController.java
@@ -9,9 +9,12 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.simp.SimpMessageSendingOperations;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 
 /** 실시간 경매 정보 처리를 위한 컨트롤러 (redis 연동) */
@@ -19,7 +22,7 @@ import java.util.List;
 @RequiredArgsConstructor
 @RestController
 public class RealTimeAuctionController {
-
+    private final SimpMessageSendingOperations simp;
     private final RealTimeAuctionService redisService;
 
     /* 경매 생성 메서드 */
@@ -73,5 +76,15 @@ public class RealTimeAuctionController {
     ) {
         Long result = redisService.finishAuctionWithBidder(auctionKey, bidderId, finalPrice);
         return ResponseEntity.ok(result);
+    }
+
+    /* 경매 좋아요 +-1 메서드 */
+    @MessageMapping("/like")
+    public void updateLike(Map<String, String> map) {
+        String auctionKey = map.get("auctionId");
+        Long memberId = Long.parseLong(map.get("memberId"));
+        Long result = redisService.updateLike(auctionKey, memberId);
+
+        simp.convertAndSend("/sub/channel/" + auctionKey, result);
     }
 }
