+++ src/main/java/com/artfolio/artfolio/dto/RealTimeAuctionPreviewRes.java
@@ -1,5 +1,6 @@
 package com.artfolio.artfolio.dto;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 public record RealTimeAuctionPreviewRes(
@@ -27,6 +28,7 @@ public record RealTimeAuctionPreviewRes(
             Long currentPrice,
             String auctionTitle,
             String artPieceTitle,
+            LocalDateTime createdAt,
             String thumbnailPath
     ) {
         public static PreviewInfo of(RealTimeAuctionInfo info) {
@@ -36,6 +38,7 @@ public record RealTimeAuctionPreviewRes(
                     info.getAuctionCurrentPrice(),
                     info.getAuctionTitle(),
                     info.getArtPieceTitle(),
+                    info.getCreatedAt(),
                     info.getPhotoPaths().size() == 0 ? "null" : info.getPhotoPaths().get(0)
             );
         }
